+++ src/main/kotlin/dev/slint/ideaplugin/ide/editor/SlintQuoteHandler.kt
@@ -1,0 +1,6 @@
+package dev.slint.ideaplugin.ide.editor
+
+import com.intellij.codeInsight.editorActions.SimpleTokenSetQuoteHandler
+import dev.slint.ideaplugin.lang.psi.SLINT_STRINGS
+
+class SlintQuoteHandler: SimpleTokenSetQuoteHandler(SLINT_STRINGS)